Move user commands to their own subfolder
