+++ src/command-line/users/index.js
@@ -1,0 +1,7 @@
+"use strict";
+
+require("./list");
+require("./add");
+require("./remove");
+require("./reset");
+require("./edit");
